+++ src/lib/s21_cos.c
@@ -12,7 +12,7 @@ long double s21_cos(double x) {
   long double ith_term = 1;  // 1 is the zeroth term
   long double partial_sum = ith_term;
 
-  for (int i = 1; s21_fabs(ith_term) > S21_EPSILON; ++i) {
+  for (int i = 1; s21_fabs((double)ith_term) > S21_EPSILON; ++i) {
     ith_term = -ith_term * x * x / ((2 * i) * (2 * i - 1));
     partial_sum += ith_term;
   }
